refactor: clean up null safety and redundant default values across modules
This commit performs a widespread refactoring to improve code conciseness and leverage Kotlin's null safety more effectively. It primarily focuses on removing redundant null-coalescing operators (`?:`) where the underlying property is already non-nullable or has a sensible default, and simplifying logic in UI components and data managers.

Specific changes include:

- **Settings Feature**:
    - Removed redundant `?: 0`, `?: false`, and `?: ""` defaults in various configuration screens (Audio, Display, Telemetry, Serial, User, Network, MQTT, etc.) where form state values are already guaranteed to be non-null.
    - Updated `DeviceConfigScreen` and `DisplayConfigScreen` to suppress deprecation warnings where necessary.
    - Simplified logic for interval dropdown selections by removing unnecessary null checks.

- **Node Feature**:
    - Refactored `TracerouteLog`, `NeighborInfoLog`, and `NodeDetailActions` to remove redundant null checks on `long_name` and `short_name`.
    - Cleaned up metric displays (Pax, Power, Signal, Device, Environment) by removing unnecessary null safety operators on timestamps and telemetry values.
    - Updated `NodeStatusIcons` to explicitly set `TooltipAnchorPosition.Above`.
    - Fixed a permission annotation in `AndroidPhoneLocationProvider`.

- **Core Modules**:
    - **Data**: Simplified property access in `NodeManagerImpl` and `MqttManagerImpl`.
    - **Model**: Removed redundant null checks in `Node`, `NodeInfo`, and `Channel` models. Updated `hasPKC` and `mismatchKey` logic for clarity.
    - **UI**: Refactored `SecurityIcon`, `ProtoExtensions`, and `EditListPreference` to streamline property access and remove unnecessary null handling.
    - **Database**: Cleaned up key existence checks in `NodeInfoDao` and simplified name matching logic.

- **Build Logic**:
    - Removed the deprecated `android.newDsl=false` property from `gradle.properties`.
    - Migrated `ExportDataUseCase` to use `kotlin.time.Instant` instead of `kotlinx.datetime.Instant`.

+++ feature/node/src/commonMain/kotlin/org/meshtastic/feature/node/component/EnvironmentMetrics.kt
@@ -167,7 +167,7 @@ internal fun EnvironmentMetrics(
                         add(
                             VectorMetricInfo(
                                 label = Res.string.wind,
-                                value = ws.toFloat().toSpeedString(displayUnits),
+                                value = ws.toSpeedString(displayUnits),
                                 icon = Icons.Outlined.Navigation,
                                 rotateIcon = normalizedBearing.toFloat(),
                             ),
